+++ packages/stormy_kit/lib/core/network/stormy_network.dart
@@ -4,6 +4,8 @@
 export 'package:dio/dio.dart'
     show
         DioException,
+        ResponseBody,
+        ResponseType,
         Response,
         Options,
         CancelToken,