+++ packages/stormy_i18n/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* Require Dart 3.11 or later.
+* Upgrade to `analyzer` 14 for compatibility with `hive_ce_generator` 1.11.3.
+* Migrate the Dart translation parser to the analyzer 13+ AST APIs.
+
 ## 0.0.1
 
 * TODO: Describe initial release.